+++ apps/web/ui/partners/lander/blocks/accordion-block.tsx
@@ -6,8 +6,8 @@ import {
   AccordionTrigger,
 } from "@dub/ui";
 import { z } from "zod";
-import { BlockMarkdown } from "./BlockMarkdown";
-import { BlockTitle } from "./BlockTitle";
+import { BlockMarkdown } from "./block-markdown";
+import { BlockTitle } from "./block-title";
 
 export function AccordionBlock({
   block,
