+++ apps/web/ui/partners/lander/blocks/block-description.tsx
