+++ apps/web/ui/partners/lander/blocks/earnings-calculator-block.tsx
@@ -7,8 +7,8 @@ import NumberFlow from "@number-flow/react";
 import { useId, useState } from "react";
 import { z } from "zod";
 import { formatRewardDescription } from "../../format-reward-description";
-import { BlockDescription } from "./BlockDescription";
-import { BlockTitle } from "./BlockTitle";
+import { BlockDescription } from "./block-description";
+import { BlockTitle } from "./block-title";
 
 export function EarningsCalculatorBlock({
   block,
