+++ apps/web/ui/partners/lander/blocks/files-block.tsx
@@ -2,7 +2,7 @@ import { ProgramProps } from "@/lib/types";
 import { programLanderFilesBlockSchema } from "@/lib/zod/schemas/program-lander";
 import { ArrowUpRight, Download } from "@dub/ui/icons";
 import { z } from "zod";
-import { BlockTitle } from "./BlockTitle";
+import { BlockTitle } from "./block-title";
 
 export function FilesBlock({
   block,
