+++ apps/web/ui/partners/lander/blocks/image-block.tsx
