+++ apps/web/ui/partners/lander/blocks/index.tsx
@@ -1,9 +1,9 @@
 import z from "@/lib/zod";
 import { programLanderBlockSchema } from "@/lib/zod/schemas/program-lander";
-import { AccordionBlock } from "@/ui/partners/lander/blocks/AccordionBlock";
-import { FilesBlock } from "@/ui/partners/lander/blocks/FilesBlock";
-import { ImageBlock } from "@/ui/partners/lander/blocks/ImageBlock";
-import { TextBlock } from "@/ui/partners/lander/blocks/TextBlock";
+import { AccordionBlock } from "@/ui/partners/lander/blocks/accordion-block";
+import { FilesBlock } from "@/ui/partners/lander/blocks/files-block";
+import { ImageBlock } from "@/ui/partners/lander/blocks/image-block";
+import { TextBlock } from "@/ui/partners/lander/blocks/text-block";
 import { EarningsCalculatorBlock } from "./earnings-calculator-block";
 
 export const BLOCK_COMPONENTS: Record<
